1.6 beta changes
Fixed misc tab inconsistencies, added relevant info onscreen when Debug is enabled, added cursorChecks to certain scripts

+++ src/main/kotlin/rat/poison/scripts/SelfNade.kt
@@ -14,7 +14,7 @@ fun selfNade() {
     Thread(Runnable {
         val angle = clientState.angle()
         angle.apply {
-            x = (-89).toDouble()
+            x = -89.0
             normalize()
         }
         clientState.setAngle(angle)
